requirements.txt
flask
openai
python-dotenv
streamlit
requests
gunicorn==21.2.0
pinecone-client